feat: Add helper function to return grid square
When the user moves the cursor over a grid square, they are actually
targeting the grid border container, rather than the grid square container.

function returnGridSquareTarget(event) does this by checking if the current
event.target is actually the square, and if not then selects the child of
the container being targeted, which would end up being the grid square itself.